Adding support for activity alert messages and summary report messages.

+++ trunk/interop/src/main/java/org/mitre/mpf/interop/JsonDetectionOutputObject.java
@@ -26,11 +26,15 @@
 
 package org.mitre.mpf.interop;
 
-import com.fasterxml.jackson.annotation.*;
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyDescription;
+import com.fasterxml.jackson.annotation.JsonTypeName;
 import org.apache.commons.lang3.ObjectUtils;
-import org.apache.commons.lang3.StringUtils;
 
-import java.util.*;
+import java.util.Objects;
+import java.util.SortedMap;
+import java.util.TreeMap;
 
 @JsonTypeName("DetectionOutputObject")
 public class JsonDetectionOutputObject implements Comparable<JsonDetectionOutputObject> {
@@ -77,8 +81,8 @@ public class JsonDetectionOutputObject implements Comparable<JsonDetectionOutput
 
 	@JsonProperty("offsetTime")
 	@JsonPropertyDescription("The offset time of this detection in the parent medium. For images, this value is not meaningful. For audio and video files, this value refers to a temporal index measured in milliseconds.")
-	private int offsetTime;
-	public int getOffsetTime() { return offsetTime; }
+	private long offsetTime;
+	public long getOffsetTime() { return offsetTime; }
 
 	@JsonProperty("artifactPath")
 	@JsonPropertyDescription("The path to the artifact containing the best representation of this detection.")
@@ -100,7 +104,7 @@ public class JsonDetectionOutputObject implements Comparable<JsonDetectionOutput
 	                                 @JsonProperty("confidence") float confidence,
 	                                 @JsonProperty("detectionProperties") SortedMap<String, String> detectionProperties,
 	                                 @JsonProperty("offsetFrame") int offsetFrame,
-									 @JsonProperty("offsetTime") int offsetTime,
+									 @JsonProperty("offsetTime") long offsetTime,
 	                                 @JsonProperty("artifactExtractionStatus") String artifactExtractionStatus,
 	                                 @JsonProperty("artifactPath") String artifactPath) {
 		this.x = x;
@@ -135,7 +139,7 @@ public class JsonDetectionOutputObject implements Comparable<JsonDetectionOutput
 		if(other == null) {
 			return 1;
 		} else if((result = Integer.compare(offsetFrame, other.offsetFrame)) != 0
-			||( result = Integer.compare(offsetTime, other.offsetTime)) != 0
+			||( result = Long.compare(offsetTime, other.offsetTime)) != 0
 			|| (result = Integer.compare(x, other.x)) != 0
 			|| (result = Integer.compare(y, other.y)) != 0
 			|| (result = Integer.compare(width, other.width)) != 0
