+++ trunk/interop/src/main/java/org/mitre/mpf/interop/JsonTrackOutputObject.java
@@ -26,11 +26,12 @@
 
 package org.mitre.mpf.interop;
 
-import com.fasterxml.jackson.annotation.*;
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyDescription;
+import com.fasterxml.jackson.annotation.JsonTypeName;
 import org.apache.commons.lang3.ObjectUtils;
-import org.apache.commons.lang3.StringUtils;
 
-import java.util.Objects;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
@@ -64,13 +65,13 @@ public class JsonTrackOutputObject implements Comparable<JsonTrackOutputObject> 
 
 	@JsonProperty("startOffsetTime")
 	@JsonPropertyDescription("The offset in the medium where the track starts, in milliseconds.")
-	private int startOffsetTime;
-	public int getStartOffsetTime() { return startOffsetTime; }
+	private long startOffsetTime;
+	public long getStartOffsetTime() { return startOffsetTime; }
 
 	@JsonProperty("stopOffsetTime")
 	@JsonPropertyDescription("The offset in the medium where the track ends, in milliseconds.")
-	private int stopOffsetTime;
-	public int getStopOffsetTime() { return stopOffsetTime; }
+	private long stopOffsetTime;
+	public long getStopOffsetTime() { return stopOffsetTime; }
 
 	@JsonProperty("type")
 	@JsonPropertyDescription("The type of object associated with this track.")
@@ -93,7 +94,7 @@ public class JsonTrackOutputObject implements Comparable<JsonTrackOutputObject> 
 	private SortedSet<JsonDetectionOutputObject> detections;
 	public SortedSet<JsonDetectionOutputObject> getDetections() { return detections; }
 
-	public JsonTrackOutputObject(String id, int startOffsetFrame, int stopOffsetFrame, int startOffsetTime, int stopOffsetTime, String type, String source) {
+	public JsonTrackOutputObject(String id, int startOffsetFrame, int stopOffsetFrame, long startOffsetTime, long stopOffsetTime, String type, String source) {
 		this.id = id;
 		this.startOffsetFrame = startOffsetFrame;
 		this.stopOffsetFrame = stopOffsetFrame;
@@ -110,8 +111,8 @@ public class JsonTrackOutputObject implements Comparable<JsonTrackOutputObject> 
 	public static JsonTrackOutputObject factory(@JsonProperty("id") String id,
 	                                            @JsonProperty("startOffsetFrame") int startOffsetFrame,
 	                                            @JsonProperty("stopOffsetFrame") int stopOffsetFrame,
-												@JsonProperty("startOffsetTime") int startOffsetTime,
-												@JsonProperty("stopOffsetTime") int stopOffsetTime,
+												@JsonProperty("startOffsetTime") long startOffsetTime,
+												@JsonProperty("stopOffsetTime") long stopOffsetTime,
 	                                            @JsonProperty("type") String type,
 	                                            @JsonProperty("source") String source,
 	                                            @JsonProperty("exemplar") JsonDetectionOutputObject exemplar,
@@ -144,8 +145,8 @@ public class JsonTrackOutputObject implements Comparable<JsonTrackOutputObject> 
 			return 0;
 		} else if((result = Integer.compare(startOffsetFrame, other.startOffsetFrame)) != 0
 				|| (result = Integer.compare(stopOffsetFrame, other.stopOffsetFrame)) != 0
-				|| (result = Integer.compare(startOffsetTime, other.startOffsetTime)) != 0
-				|| (result = Integer.compare(stopOffsetTime, other.stopOffsetTime)) != 0
+				|| (result = Long.compare(startOffsetTime, other.startOffsetTime)) != 0
+				|| (result = Long.compare(stopOffsetTime, other.stopOffsetTime)) != 0
 				|| (result = ObjectUtils.compare(type, other.type, false)) != 0
 			    || (result = ObjectUtils.compare(source, other.source, false)) != 0
 				|| (result = ObjectUtils.compare(exemplar, other.getExemplar(), false)) != 0
